feat: 修复自定义指标中文维度展示为 base62 编码 --story=137662640
